ai_ml_trading/03_rl_market_making/requirements.txt
# Core dependencies for RL Market Making System

# Core scientific computing
numpy>=1.21.0
pandas>=1.4.0
scipy>=1.8.0
scikit-learn>=1.0.0

# Deep learning and RL
torch>=1.12.0
torchvision>=0.13.0
gymnasium>=0.26.0  # Updated OpenAI Gym
stable-baselines3>=1.6.0

# Quantitative finance
quantlib>=1.26
ta-lib>=0.4.24
yfinance>=0.1.70
vectorbt>=0.25.0

# Data handling and analysis
h5py>=3.6.0
tables>=3.7.0
joblib>=1.1.0
numba>=0.56.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.8.0
bokeh>=2.4.0

# Web and API
requests>=2.27.0
aiohttp>=3.8.0
websockets>=10.0

# Configuration and logging
pyyaml>=6.0
toml>=0.10.2
python-dotenv>=0.19.0

# Experiment tracking and monitoring
wandb>=0.12.0
mlflow>=1.24.0
tensorboard>=2.8.0

# Testing and quality
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-benchmark>=3.4.0
pytest-mock>=3.6.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.942
isort>=5.10.0
pre-commit>=2.17.0

# Performance and optimization
cython>=0.29.0
numexpr>=2.8.0
bottleneck>=1.3.0

# Database and storage
sqlalchemy>=1.4.0
redis>=4.1.0
pymongo>=4.0.0

# Networking and distributed computing
zmq>=22.3.0
mpi4py>=3.1.0

# Time series and statistics
statsmodels>=0.13.0
arch>=5.0.0
pymc3>=3.11.0

# Jupyter and development
jupyter>=1.0.0
jupyterlab>=3.3.0
ipywidgets>=7.6.0
notebook>=6.4.0

# Optional crypto market data
ccxt>=1.80.0

# System monitoring
psutil>=5.8.0
memory-profiler>=0.60.0

# Miscellaneous utilities
tqdm>=4.62.0
click>=8.0.0
rich>=12.0.0
typer>=0.4.0